core: decouple gatt serial from connection events
This is a big win for seperation of concerns. Previously the Gatt Serial
profile was responsible for firing connection related callbacks which is
now handled by GattClient. As a biproduct of this change, it was easy
to implement a very important feature which is waiting for all services
to be discovered before firing onConnected().